lab2

+++ SchoolRegister.Model/DataModels/GradeScale.cs
@@ -1,10 +1,10 @@
 namespace SchoolRegister.Model.DataModels
 {
-    public enum GradeScale :float
+    public enum GradeScale :int
     {
-        NDST = 2.0,
-        DST = 3.0,
-        DB = 4.0,
-        BDB = 5.0
+        NDST = 2,
+        DST = 3,
+        DB = 4,
+        BDB = 5
     }
 }
